+++ src/components/Button/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Button';
-export * from './Button';
+export { default as Button} from './Button';
