+++ src/components/Card/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Card';
-export * from './Card';
+export { default as Card } from './Card';
