+++ src/components/Header/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Header';
-export * from './Header';
+export { default as Header } from './Header';
