+++ src/components/Icons/CheckIcon/index.ts
@@ -1,1 +1,1 @@
-export { default } from './CheckIcon';
+export { default as CheckIcon} from './CheckIcon';
