+++ src/components/Input/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Input';
-export * from './Input';
+export { default as Input } from './Input';
