+++ src/components/Loader/index.ts
@@ -1,2 +1,1 @@
-export { default } from './Loader';
-export * from './Loader';
+export { default as Loader } from './Loader';
