+++ src/components/index.ts
@@ -1,0 +1,10 @@
+export * from './Button';
+export * from './Card';
+export * from './Header';
+export * from './Input';
+export * from './Loader';
+export * from './Placeholder';
+export * from './Text';
+export * from './Icons/ArrowDownIcon';
+export * from './Icons/CheckIcon';
+export * from './CardsContainer'
